fix: resolve all clippy warnings
- Remove unnecessary type casts and closures
- Simplify Option handling with is_some_and() and map()
- Use struct initialization with ..Default::default()
- Add #[allow(clippy::too_many_arguments)] where appropriate
- Fix reference patterns and iterator usage
- Collapse nested if statements
- Use strip_prefix instead of manual slicing

+++ src/lib/graph/bfs.rs
@@ -18,7 +18,7 @@ fn check_min_value_filter(
 ) -> bool {
     if let Some(meta) = provider.get_metadata(note_id) {
         let value = meta.value.unwrap_or(50);
-        min_value.is_none_or(|min| value as u8 >= min)
+        min_value.is_none_or(|min| value >= min)
     } else {
         false
     }
